Fixed bug with the jsx compiler option not being handled properly.
The jsx compiler option was previously being passed through to tsc as a string - but tsc expects it to be an enum value of type `JsxEmit`. This change resolves the `jsx` string value to the equivalent `JsxEmit` enum value in the same way that the `module` and `target` compiler options are handled.