Uses built-in transpilation API from Typescript compiler
Rather than reimplementing the transpileModule function, use it directly. Trying to create a limited CompilerHost to feed to the tooling means that specifying alternate targets besides ES5 does not function correctly. For example, prior to this fix, even if tsconfig.json specified "target": "es2015" to include a native Promise implementation, this plugin would error out on async/await constructs because the relevant lib.d.ts files could not be resolved by the CompilerHost.